// src/app/model/user.model.ts
import {s} from '@angular/core/src/render3';

export class  User {

  password: string;
  SubPractice: string;
  PracticeId: string;
  Email1: string;
  LoginId: string;
  WorkPhoneNo: string;
  MintRole: string;
  LName: string;
  TitleTypeId: string;
  AliasName: string;
  SkillSet: string;
  IsTeamLead: string;
  OrgId: string;
  Image: string;
  CellPhoneNo: string;
  DepartmentId: string;
  IsManager: string;
  CurStatus: string;
  Location: string;
  CoverPic: string;
  Country: string;
  ResultString: string;
  FName: string;
  MName: string;
  EmployeeTypeId: string;
  IsAuthenticate: string;
  HireDate: string;
  Id: string;
  IsOperationContactTeam: string;
  AboutMe: string;
}